+++ check_args.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   check_args.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mspasic <mspasic@student.hive.fi>          +#+  +:+       +#+        */
+/*   By: aheinane <aheinane@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/18 15:02:19 by aheinane          #+#    #+#             */
-/*   Updated: 2024/10/08 17:43:20 by mspasic          ###   ########.fr       */
+/*   Updated: 2024/10/21 09:07:30 by aheinane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
